# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API for Peminjaman Laboratorium Terpadu Poltekkes Tasikmalaya
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, mahasiswa, plp, gudang, dosen]
        - name: jurusanId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{id}/verify:
    put:
      operationId: verifyUser
      tags: [users]
      summary: Verify user registration (by PLP)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyUserRequest"
      responses:
        "200":
          description: User verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /bahan:
    get:
      operationId: getBahan
      tags: [bahan]
      summary: Get all materials/supplies
      parameters:
        - name: laboratoriumId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bahan"

    post:
      operationId: createBahan
      tags: [bahan]
      summary: Create new material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBahanRequest"
      responses:
        "201":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bahan"
  /bahan/{id}:
    put:
      operationId: updateBahan
      tags: [bahan]
      summary: Update material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBahanRequest"
      responses:
        "200":
          description: Material updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bahan"

    delete:
      operationId: deleteBahan
      tags: [bahan]
      summary: Delete material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /permintaan-bahan:
    get:
      operationId: getPermintaanBahan
      tags: [permintaan-bahan]
      summary: Get material request records
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [menunggu, disetujui, ditolak, disiapkan]
        - name: userId
          in: query
          schema:
            type: integer
        - name: tanggalMulai
          in: query
          schema:
            type: string
            format: date
        - name: tanggalSelesai
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of material requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermintaanBahan"

    post:
      operationId: createPermintaanBahan
      tags: [permintaan-bahan]
      summary: Create material request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePermintaanBahanRequest"
      responses:
        "201":
          description: Material request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermintaanBahan"
  /permintaan-bahan/{id}:
    get:
      operationId: getPermintaanBahanById
      tags: [permintaan-bahan]
      summary: Get material request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermintaanBahan"
  /permintaan-bahan/{id}/status:
    put:
      operationId: updatePermintaanBahanStatus
      tags: [permintaan-bahan]
      summary: Update material request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermintaanBahan"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    VerifyUserRequest:
      type: object
      properties:
        status:
          type: string
          enum: [aktif, ditolak]
        catatan:
          type: string
      required: [status]
    User:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, mahasiswa, plp, gudang, dosen]
        nim:
          type: string
          nullable: true
        nip:
          type: string
          nullable: true
        noHp:
          type: string
          nullable: true
        jurusanId:
          type: integer
          nullable: true
        jurusan:
          $ref: "#/components/schemas/Jurusan"
          nullable: true
        laboratoriumId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [menunggu, aktif, nonaktif, ditolak]
        createdAt:
          type: string
          format: date-time
      required: [id, nama, email, role, status, createdAt]
    CreateUserRequest:
      type: object
      properties:
        nama:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, mahasiswa, plp, gudang, dosen]
        nim:
          type: string
          nullable: true
        nip:
          type: string
          nullable: true
        noHp:
          type: string
          nullable: true
        jurusanId:
          type: integer
          nullable: true
        laboratoriumId:
          type: integer
          nullable: true
      required: [nama, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        nama:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, mahasiswa, plp, gudang, dosen]
        nim:
          type: string
          nullable: true
        nip:
          type: string
          nullable: true
        noHp:
          type: string
          nullable: true
        jurusanId:
          type: integer
          nullable: true
        laboratoriumId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [menunggu, aktif, nonaktif, ditolak]
    Jurusan:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        kode:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, nama, kode, createdAt]
...
    Laboratorium:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        kode:
          type: string
        lokasi:
          type: string
        kapasitas:
          type: integer
        jurusanId:
          type: integer
          nullable: true
        jurusan:
          $ref: "#/components/schemas/Jurusan"
          nullable: true
        deskripsi:
          type: string
          nullable: true
        fasilitas:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, nama, kode, lokasi, kapasitas, createdAt]
...
    Bahan:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        kode:
          type: string
        deskripsi:
          type: string
          nullable: true
        stok:
          type: integer
        stokMinimal:
          type: integer
        satuan:
          type: string
        laboratoriumId:
          type: integer
        laboratorium:
          $ref: "#/components/schemas/Laboratorium"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, nama, kode, stok, stokMinimal, satuan, laboratoriumId, createdAt]
    CreateBahanRequest:
      type: object
      properties:
        nama:
          type: string
        kode:
          type: string
        deskripsi:
          type: string
          nullable: true
        stok:
          type: integer
        stokMinimal:
          type: integer
        satuan:
          type: string
        laboratoriumId:
          type: integer
      required: [nama, kode, stok, stokMinimal, satuan, laboratoriumId]
...
    PermintaanBahanItem:
      type: object
      properties:
        bahanId:
          type: integer
        bahan:
          $ref: "#/components/schemas/Bahan"
          nullable: true
        jumlahDiminta:
          type: integer
        jumlahDisetujui:
          type: integer
          nullable: true
      required: [bahanId, jumlahDiminta]
    PermintaanBahan:
      type: object
      properties:
        id:
          type: integer
        noPermintaan:
          type: string
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        laboratoriumId:
          type: integer
          nullable: true
        tanggalDibutuhkan:
          type: string
          format: date
        keperluan:
          type: string
        status:
          type: string
          enum: [menunggu, disetujui, ditolak, disiapkan]
        catatanGudang:
          type: string
          nullable: true
        verifikasiOleh:
          type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PermintaanBahanItem"
        createdAt:
          type: string
          format: date-time
      required: [id, noPermintaan, userId, tanggalDibutuhkan, keperluan, status, items, createdAt]
    CreatePermintaanBahanRequest:
      type: object
      properties:
        laboratoriumId:
          type: integer
          nullable: true
        tanggalDibutuhkan:
          type: string
          format: date
        keperluan:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              bahanId:
                type: integer
              jumlahDiminta:
                type: integer
            required: [bahanId, jumlahDiminta]
      required: [tanggalDibutuhkan, keperluan, items]
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
        catatan:
          type: string
          nullable: true
        jumlahDisetujui:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: integer
              jumlah:
                type: integer
          nullable: true
      required: [status]
...